requirements.txt
flask
line-bot-sdk
configparser
vertexai